==> python_data_analytics/pandas/00-pandas_basics.ipynb <==
# Pandas
<p>Pandas is a popular open-source data manipulation and analysis library for the python programming language.  It provides a powerful and flexible set of tools for working with structured data, making it a fundamental tool for data scientists, analysts and engineers.</p>
<p>Pandas is esigned to handle data in various formats, such as tabular data, time series data and more making it an essentail part of the data processing workflow in many industries.</p>
<h5>Key Features and Functionalities of Pandas</h5>
<ul>
    <li>ofers two primary data structures;<em>DataFrame</em> and <em>Series</em></li>
    <li>makes it easy to read data from various sources, including CSV files, excel spreadsheets, SQL databases and more. It can also export data to these formats, enabling seamless data exchange</li>
    <li>Data Mergin and Joining: Multiple DataFrames can be combined using methods like merge and join, similar to SQL operations, to create more complex datasets from different sources.</li>
    <li>provides efficient indexing and selection methods, allowing for quick access to specific rows and columns of data</li>
    <li>used to create custom data structures and manipulate date in ways that suit specific needs, extending Panda's capabilities.</li>
</ul>
<p></p>

We begin by installing pandas

!pip install pandas

Then we import into our python file

import pandas as pd

# create a list
column = ['Nigeria', 'E. Guinea', 'Ivory Coast', 'G. Bissau']

print(column)

Whenever we print a python list all we get is a plain text with no special structure.

# Lets take the list and insert into a dataframe
data = pd.DataFrame(column)
print(data)

with pandas, we can organize our list/data into rows and columns thereby giving it a structure and making it more easier to view.

# name the column
column_title = {"Team": column}
data = pd.DataFrame(column_title)

print(data)

# add new columns to our dataframe
column_title = {"Team": column, 
                "Points": [7,7,3,0], 
                "GD": [2,6,-3,-5], 
                "Won": [2,2,1,0],
               "Lost":[0,0,2,3],
               "Draw":[1,1,0,0]}
data = pd.DataFrame(column_title)

print(data)

# select values from our dataframe
select_column = data['Points']
print(select_column)

# to select a specific value we add index to the key
# Select 'Ivory Coasts' goal difference
select_column = data['GD'][2]
print(select_column)


# select the entire row for Ivory Coast
select_row = data.iloc[2]

print(select_row)

# to select specific value using this method
select_row = data.iloc[2]['GD']

print(select_row)

<h5>Class Exercise:</h5>
<p>Create a dataframe for name, age and height of 4 individuals</p>

# Solution
records = pd.DataFrame({"Name": ['juliet', 'Leonard', 'sophia', 'Daniel'],
                       "Weight": [50, 60, 58, 68],
                       "Height": [1.65, 1.73, 1.66, 1.69]})

print(records)

bmi = []
#weight/(height**2)
for i in range(len(records)):
    bmi_score = records['Weight'][i]/(records['Height'][i]**2)
    bmi.append(bmi_score)
    
#create a new column to assign BMI scores
records['bmi'] = bmi

print(records)


# save dataframe into a file
records.to_csv('bmi.csv', index=False)

# we can use the sep to choose delimeters
# records.to_csv('bmi.csv', sep="\t")

# we can also save to txt
data.to_csv('group_A.txt', index=False, sep="\t")

# to get ride of unnessary column we specify index to be false
# data.to_csv('group_A.txt', index=False, sep="\t")
