==> python_data_analytics/pandas/01-pandas_load.ipynb <==
## Loading data from an existing file

import pandas as pd
import sqlite3

data = pd.read_csv('bmi.csv')
print(data)

# loading the tsv file
data = pd.read_csv("group_A.txt", sep='\t')
print(data)

## Loading data from a database

# connect to the database
connection = sqlite3.connect("fifa.db")

gta_data = pd.read_sql("select * from fifa_records", connection)
print(gta_data)

forbes = pd.read_csv('forbes_2022.csv', sep=",")
#print(forbes)

# to print the first 5 rows
# print(forbes.head())
# we can also specify the number of rows from the beginning
# print(forbes.head(10))


# To print the last 5 rows
print(forbes.tail())

# filter our data entry
filtered_row = forbes[forbes['category'] == 'Automotive']
#print(filtered_row)

# to print specific rows
print(forbes['category'].head())

# to replace every instance of
replaced_category = forbes.replace('Automotive', 'Automobile')
# print(replaced_category)

# remove some data // remove the about
remove_column = forbes.drop("about", axis=1)
# print(remove_column)

# to remove more than one column
remove_column = forbes.drop(['about', 'bio'], axis=1)
# print(remove_column)

# to get rid of rows // delete rows
remove_row = forbes.iloc[1:4]
# print(remove_row)

# add new rows to our database
row = {"Team":"Ghana", "Points": 0, "GD": -5, "Won": 0, "Lost": 2, "Draw": 1}
new_data = data.append(row, ignore_index=True)
# print(new_data)
