==> python_data_analytics/pandas/02-drop_duplicates.ipynb <==
import pandas as pd

jam_data = pd.read_csv("jamData.csv")
print("number of participants:", len(jam_data))

# get rid of the BOT data
jam_data = jam_data.drop_duplicates(subset=['name / nickname (optional)'])
print("number of participants:", len(jam_data))
