==> python_data_analytics/pandas/03-scaper_pandas.ipynb <==
import pandas as pd
import ssl



ssl._create_default_https_context = ssl._create_unverified_context

scraper = pd.read_html("https://en.wikipedia.org/wiki/Comparison_of_Linux_distributions")
scraper

# to view all tables
for idx, table in enumerate(scraper):
    print("**************************************")
    print(idx)
    print(table)

# select the particular table
scraper[3]
