==> python_data_analytics/pandas/04-sqlite_database.ipynb <==
# import sqlite3
import sqlite3


# create an empty sqlite database
connection = sqlite3.connect("gta.db")

# create a cursor object to be in charge of all our communication with the database
cursor = connection.cursor()

# create a table
cursor.execute("Create table gta (release_year integer, release_name text, city text)")

release_list = [
                (1997, "Grand Theft Auto", "State of New Guernsey"),
               (1999, "Grand Theft Auto 2", "Anywhere, USA"),
               (2001, "Grand Theft Auto III", "Liberty City"),
               (2002, "Grand Theft Auto: Vice City", "Vice City"),
               (2004, "Grand Theft Auto", "State of San Andreas"),
               (2008, "Grand Theft Auto", "Liberty City"),
               (2013, "Grand Theft Auto V", "Los Santos")
               ]

cursor.executemany("insert into gta values (?,?,?)", release_list)

#print database rows
for row in cursor.execute("Select * from gta"):
    print(row)

# terminate established connection
connection.close()

# Fifa World Cup Winners

# create an empty sqlite database
connection = sqlite3.connect("fifa.db")

# create a cursor object to be in charge of all our communication with the database
cursor = connection.cursor()

# create a table
cursor.execute("Create table fifa_records (year integer, winner text, location text, attendace integer)")

fifa_list = [
                (1994, "Brazil", "California, United States", 94194),
               (1998, "France", "Saint-Denis, France", 80000),
               (2002, "Brazil", "Berlin, Germany", 69029),
               (2006, "Italy", "California, United States", 69000),
              (2010, "Spain", "Johannesburg, South Africa", 84490),
              (2014, "Germany", "Rio de janeiro, Brazil", 74738),
            (2018, "France", "Moscow, Russia", 78011),
            (2022, "Argentina", "Lusail, Qatar", 88966)
               ]

cursor.executemany("insert into fifa_records values (?,?,?,?)", fifa_list)

#print database rows
for row in cursor.execute("Select * from fifa_records"):
    print(row)

# terminate established connection
connection.close()
