==> python_data_analytics/pandas/05-forbes.ipynb <==
## Working with the forbes list

# import pandas
import pandas as pd

# load forbes data set
forbesList = pd.read_csv("forbes_2022.csv", sep=",")
print(forbesList)

print(forbesList.head(2))

print(forbesList.tail())

# filter data
filtered_row = forbesList[forbesList["category"] == "Automotive"]
print(filtered_row)

# replace values in data set
replaced_category = forbesList.replace("Automotive", "Automobile")

print(replaced_category.head())

# remove data from dataframe 
remove_column = forbesList.drop('numberOfSiblings', axis=1)
print(remove_column.head(2))

new_forbes = forbesList.drop(['year','month', 'source', 'state', 'country', 'state', 'organization', 'philanthropyScore',
                             'birthDate', 'residenceMsa', 'bio', 'about', 'title', 'numberOfSiblings', 'city' ], axis=1)

print(new_forbes.head())

print(new_forbes)

# get unique values in a column
countries = new_forbes["countryOfCitizenship"].unique()
print(countries)
print(len(countries))

europeSrting = """Albania
Andorra
Armenia
Austria
Azerbaijan
Belarus
Belgium
Bosniaand Herzegovina
Bulgaria
Croatia
Cyprus
Czechia
Denmark
Estonia
Finland
France
Georgia
Germany
Greece
Hungary
Iceland
Ireland
Italy
Kazakhstan
Kosovo
Latvia
Liechtenstein
Lithuania
Luxembourg
Malta
Moldova
Monaco
Montenegro
Netherlands
North Macedonia
Norway
Poland
Portugal
Romania
Russia
San Marino
Serbia
Slovakia
Slovenia
Spain
Sweden
Switzerland
Turkey
Ukraine
United Kingdom
VaticanCity
"""

europe = europeSrting.splitlines()
print(europe)

asiaString = """Afghanistan
Armenia
Azerbaijan
Bahrain
Bangladesh
Bhutan
Brunei
Cambodia
China
Cyprus
Georgia
India
Indonesia
Iran
Iraq
Israel
Japan
Jordan
Kazakhstan
Kuwait
Kyrgyzstan
Laos
Lebanon
Malaysia
Maldives
Mongolia
Myanmar
Nepal
North Korea
Oman
Pakistan
Palestine
Philippines
Qatar
Russia
Saudi Arabia
Singapore
South Korea
Sri Lanka
Syria
Taiwan
Tajikistan
Thailand
Timor-Leste
Turkey
Turkmenistan
United Arab Emirates (UAE)
Uzbekistan
Vietnam
Yemen
"""

asia = asiaString.splitlines()
print(asia)

africaString = """
Algeria
Angola
Benin
Botswana
Burkina Faso
Burundi
Cabo Verde
Cameroon
Central African Republic
Chad
Comoros
Congo, Democratic Republic of the
Congo, Republic of the
Cote d'Ivoire
Djibouti
Egypt
Equatorial Guinea
Eritrea
Eswatini
Ethiopia
Gabon
Gambia
Ghana
Guinea
Guinea-Bissau
Kenya
Lesotho
Liberia
Libya
Madagascar
Malawi
Mali
Mauritania
Mauritius
Morocco
Mozambique
Namibia
Niger
Nigeria
"""

africa = africaString.splitlines()
print(africa)

northAmericaString = """Antigua and Barbuda
Bahamas
Barbados
Belize
Canada
Costa Rica
Cuba
Dominica
Dominican Republic
El Salvador
Grenada
Guatemala
Haiti
Honduras
Jamaica
Mexico
Nicaragua
Panama
Saint Kitts and Nevis
Saint Lucia
Saint Vincent and the Grenadines
Trinidad and Tobago
United States
"""
northAmerica = northAmericaString.splitlines()
print(northAmerica)

southAmericaString = """Argentina
Bolivia
Brazil
Chile
Colombia
Ecuador
Guyana
Paraguay
Peru
Suriname
Uruguay
Venezuela
"""

southAmerica = southAmericaString.splitlines()
print(southAmerica)

australiaString = """Australia
Fiji
Kiribati
Marshall Islands
Micronesia
Nauru
New Zealand
Palau
Papua New Guinea
Samoa
Solomon Islands
Tonga
Tuvalu
Vanuatu
"""

australia = australiaString.splitlines()
print(australia)

###

continent = []

for i in range(len(new_forbes)):
    if new_forbes["countryOfCitizenship"][i] in africa:
        add = "Africa"
    elif new_forbes["countryOfCitizenship"][i] in europe:
        add = "Europe"
    elif new_forbes["countryOfCitizenship"][i] in asia:
        add = "Asia"
    elif new_forbes["countryOfCitizenship"][i] in northAmerica:
        add = "North America"
    elif new_forbes["countryOfCitizenship"][i] in southAmerica:
        add = "South America"
    elif new_forbes["countryOfCitizenship"][i] in australia:
        add = "Australia"
    else:
        add = None
    continent.append(add)
    
#create a new column to assign BMI scores
new_forbes['continent'] = continent

print(new_forbes)


age_classification = []
#weight/(height**2)
for i in range(len(new_forbes)):
    if new_forbes["age"][i] < 45:
        result = 'youth'
    elif new_forbes["age"][i] < 61:
        result = 'middle age'
    elif new_forbes["age"][i] < 101:
        result = 'old age'
    else:
        result = 'NA'
        
    age_classification.append(result)
    
#create a new column to assign BMI scores
new_forbes['Age Classification'] = age_classification

print(new_forbes)


new_forbes.to_csv("forbesList.csv", index=False)
